move the highjack functions to util.py
