+++ Dolphin scripts/Entrance Randomizer/lib/utils.py
@@ -7,6 +7,12 @@ from typing import ClassVar
 from dolphin import gui  # pyright: ignore[reportMissingModuleSource]
 from lib.constants import *  # noqa: F403
 from lib.constants import __version__
+from lib.entrance_rando import (
+    temples,
+    Transition,
+    TRANSITION_INFOS_DICT,
+    transitions_map,
+)
 from lib.types_ import SeedType
 
 DRAW_TEXT_STEP = 24
@@ -60,6 +66,67 @@ def follow_pointer_path(ppath: Sequence[int]):
                 + str([hex(level).upper() for level in ppath]),
             )
     return addr
+
+
+def highjack_transition(
+    from_: int | None,
+    to: int | None,
+    redirect: int,
+):
+    if from_ is None:
+        from_ = state.current_area_old
+    if to is None:
+        to = state.current_area_new
+
+    # Early return. Detect the start of a transition
+    if state.current_area_old == state.current_area_new:
+        return False
+
+    if from_ == state.current_area_old and to == state.current_area_new:
+        print(
+            "highjack_transition |",
+            f"From: {hex(state.current_area_old)},",
+            f"To: {hex(state.current_area_new)}.",
+            f"Redirecting to: {hex(redirect)}",
+        )
+        memory.write_u32(ADDRESSES.current_area, redirect)
+        return True
+    return False
+
+
+def highjack_transition_rando():
+    # Early return, faster check. Detect the start of a transition
+    if state.current_area_old == state.current_area_new:
+        return False
+
+    redirect = transitions_map.get((state.current_area_old, state.current_area_new))
+    if not redirect:
+        return False
+
+    # Apply Altar of Ages logic to St. Claire's Excavation Camp
+    if redirect.to in {LevelCRC.ST_CLAIRE_DAY, LevelCRC.ST_CLAIRE_NIGHT}:
+        redirect = Transition(
+            redirect.from_,
+            to=LevelCRC.ST_CLAIRE_NIGHT if state.visited_altar_of_ages else LevelCRC.ST_CLAIRE_DAY,
+        )
+
+    # Check if you're visiting a Temple for the first time, if so go directly to Spirit Fight
+    if redirect.to in temples:
+        spirit = TRANSITION_INFOS_DICT[redirect.to].exits[1].area_id
+        if not state.visited_spirits[spirit]:
+            redirect = Transition(from_=redirect.to, to=spirit)
+
+    print(
+        "highjack_transition_rando |",
+        f"From: {hex(state.current_area_old)},",
+        f"To: {hex(state.current_area_new)}.",
+        f"Redirecting to: {hex(redirect.to)}",
+        f"({hex(redirect.from_)} entrance)\n",
+    )
+    memory.write_u32(follow_pointer_path(ADDRESSES.prev_area), redirect.from_)
+    memory.write_u32(ADDRESSES.current_area, redirect.to)
+    state.current_area_new = redirect[1]
+    return redirect
 
 
 def prevent_transition_softlocks():
